requirements.txt
numpy==1.18.5
streamlit==0.85.1
opencv-python-headless
Pillow==7.2.0
tensorflow==2.3.1
wget==3.2